+++ generate_package.py
@@ -23,14 +23,16 @@ def build_recursive_file_list(dir_path):
             return_list.extend(build_recursive_file_list(full_path))
     return return_list
 
+
 def remove_root_path(in_path):
     last_bit = in_path.split("no_mans_sky_base_builder")[-1]
     path = "no_mans_sky_base_builder/" + last_bit
     return path
 
+
 this_dir = os.path.dirname(os.path.realpath(__file__))
 zip_out = os.path.join(this_dir, "no_mans_sky_base_builder.zip")
-with ZipFile(zip_out, 'w', zipfile.ZIP_DEFLATED) as zip_obj:
+with ZipFile(zip_out, "w", zipfile.ZIP_DEFLATED) as zip_obj:
     # Add multiple files to the zip
     for item in build_recursive_file_list(python_package_dir):
         zip_obj.write(item, remove_root_path(item))